% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{BIC.tsissm.estimate}
\alias{BIC.tsissm.estimate}
\alias{BIC}
\alias{BIC.tsissm.selection}
\title{Bayesian Information Criterion}
\usage{
\method{BIC}{tsissm.estimate}(object, ...)

\method{BIC}{tsissm.selection}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsissm.estimate}.}

\item{...}{not currently used.}
}
\value{
A numeric value.
}
\description{
Extract the BIC from an estimated model.
}
