% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{estfun.tsissm.estimate}
\alias{estfun.tsissm.estimate}
\title{Score Method}
\usage{
\method{estfun}{tsissm.estimate}(x, ...)
}
\arguments{
\item{x}{an object of class \dQuote{tsissm.estimate}.}

\item{...}{not currently used.}
}
\value{
The score matrix
}
\description{
Score Method
}
\details{
The function returns the scores of the negative of the log likelihood
at the optimal solution.
}
\author{
Alexios Galanos
}
