% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maunaloa}
\alias{maunaloa}
\title{Mauna Loa CO2 Daily Measurements}
\format{
## `maunaloa`
A data.frame containing 15571 observations
\describe{
  \item{date}{The YYYY-MM-DD date}
  \item{co2}{The CO2 measurements}
}
}
\source{
\url{https://gml.noaa.gov/ccgg/trends/data.html}
}
\usage{
maunaloa
}
\description{
The dataset represents the daily average of the number of carbon dioxide 
molecules in a given number of molecules of air, after removal of water 
vapor (molefrac) from the Mauna Loa observatory, spanning the period
1974-05-19 to 2025-02-25.
}
\keyword{datasets}
