% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nobs.tsissm.estimate}
\alias{nobs.tsissm.estimate}
\alias{nobs}
\title{Extract the Number of Observations}
\usage{
\method{nobs}{tsissm.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsissm.estimate}.}

\item{...}{not currently used.}
}
\value{
A numeric value.
}
\description{
Extract the number of observations from an estimated model.
This is principally intended to be used in computing BIC and used in other
tidy methods
}
