% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdiagnose.R
\name{print.tsissm.diagnose}
\alias{print.tsissm.diagnose}
\title{Model Diagnostics Print method}
\usage{
\method{print}{tsissm.diagnose}(x, ...)
}
\arguments{
\item{x}{an object of class \dQuote{tsissm.duagnose} generated from
calling \code{\link[tsissm]{tsdiagnose}}.}

\item{...}{not currently used.}
}
\value{
Invisibly returns the original object and prints the output to console.
}
\description{
Print method for class \dQuote{tsissm.diagnose}
}
