% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sigma.tsissm.estimate}
\alias{sigma.tsissm.estimate}
\alias{sigma}
\title{The Standard Deviation of the model}
\usage{
\method{sigma}{tsissm.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsissm.estimate}.}

\item{...}{none.}
}
\value{
If the model was estimated using dynamic variance (GARCH model), then
an xts matrix is returned with the estimated GARCH volatility else the estimated
standard deviation of the residuals in the case of constant variance.
}
\description{
The Standard Deviation of the model
}
