% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spy}
\alias{spy}
\title{SPY ETF Adjusted Close}
\format{
## `spy`
A data.frame containing 7597 observations
\describe{
  \item{date}{The YYYY-MM-DD date}
  \item{spy}{The adjusted stock closing price.}
}
}
\source{
From Yahoo Finance.
}
\usage{
spy
}
\description{
The adjusted daily closing price of the SPY ETF.
}
\keyword{datasets}
