% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.tsissm.estimate}
\alias{summary.tsissm.estimate}
\alias{summary}
\title{Model Estimation Summary}
\usage{
\method{summary}{tsissm.estimate}(object, vcov_type = "H", digits = 4, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsissm.estimate}.}

\item{vcov_type}{the type of standard errors based on the vcov estimate (see \code{\link{vcov}}).}

\item{digits}{integer, used for number formatting. Optionally, to avoid
scientific notation, set \sQuote{options(scipen=999)}.}

\item{...}{not currently used.}
}
\value{
A printout of the parameter summary, model type and some model metrics.
}
\description{
Summary method for class \dQuote{tsissm.estimate}
}
