% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdiagnose.R
\name{tsdiagnose.tsissm.estimate}
\alias{tsdiagnose.tsissm.estimate}
\alias{tsdiagnose}
\title{Model Diagnostics}
\usage{
\method{tsdiagnose}{tsissm.estimate}(object, plot = FALSE, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsissm.estimate}.}

\item{plot}{whether to generate diagnostic plots to accompany summary.}

\item{...}{not currently used.}
}
\value{
A list of tests including the weighted Ljung-Box test for residual 
autocorrelation, system forecastability test, and outlier dates using the 
Rosner test (\code{\link[EnvStats]{rosnerTest}}). Optionally generates a plot
of relevant diagnostics.
}
\description{
Creates a short summary of model based diagnostics.
}
