% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsequation.tsissm.estimate}
\alias{tsequation.tsissm.estimate}
\alias{tsequation}
\title{Model Equation (LaTeX)}
\usage{
\method{tsequation}{tsissm.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsissm.estimate}.}

\item{...}{not currently used.}
}
\value{
A list of equations in LaTeX which can be used in documents. This is
a list with 3 slots for the observation, state and variance equations,
}
\description{
Generates a list of model equations in LaTeX.
}
\details{
This method is called in the summary when the format output option
chosen is \dQuote{flextable}.
}
