% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsmetrics.tsissm.predict}
\alias{tsmetrics.tsissm.predict}
\alias{tsmetrics}
\alias{tsmetrics.tsissm.estimate}
\title{Performance Metrics}
\usage{
\method{tsmetrics}{tsissm.predict}(object, actual, alpha = 0.1, ...)

\method{tsmetrics}{tsissm.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsissm.estimate} or \dQuote{tsissm.predict}}

\item{actual}{the actual data matched to the dates of the forecasts.}

\item{alpha}{the coverage level for distributional forecast metrics.}

\item{...}{not currently used.}
}
\value{
a data.frame of performance metrics including MAPE, MSLRE, BIAS and AIC
for the estimate object and MAPE, MSLRE, BIAS, MASE, MIS and CRPS for predict object.
}
\description{
Performance metrics from an estimated or predicted tsissm model.
}
