% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{tsspec.tsissm.estimate}
\alias{tsspec.tsissm.estimate}
\alias{tsspec}
\title{Model Specification Extractor}
\usage{
\method{tsspec}{tsissm.estimate}(
  object,
  y = NULL,
  lambda = NULL,
  xreg = NULL,
  parameters = "initial",
  ...
)
}
\arguments{
\item{object}{an object of class \dQuote{tsissm.estimate}.}

\item{y}{an optional new xts vector.}

\item{lambda}{an optional lambda parameter for the Box Cox transformation (if
previously used).}

\item{xreg}{an optional matrix of regressors.}

\item{parameters}{whether to set the parameters to their initial values, \dQuote{initial}
or the optimally estimated \dQuote{optimal}.}

\item{...}{not currently used.}
}
\value{
An object of class \dQuote{tsissm.spec}.
}
\description{
Extracts a model specification (class \dQuote{tsissm.spec}) from
an object of class \dQuote{tsissm.estimate}.
}
\note{
This function is used by other functions in the package such as the
backtest which requires rebuilding the specification for each re-estimation
step with updated data but keeping all else equal.
}
