% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_retail_sales}
\alias{us_retail_sales}
\title{Advance Retail Sales: Retail Trade (US)}
\format{
## `us_retail_sales`
A data.frame containing 398 observations
\describe{
  \item{date}{The YYYY-MM-DD date}
  \item{co2}{The sales value in millions of dollars.}
}
}
\source{
U.S. Census Bureau, Advance Retail Sales: Retail Trade [RSXFSN]. Retrieved from FRED, Federal Reserve Bank of St. Louis: \url{https://fred.stlouisfed.org/series/RSXFSN} on March 27, 2025.
}
\usage{
us_retail_sales
}
\description{
The dataset represents the monthly, non-seasonally adjusted advance estimate
of sales from the retail trade based on data from a subsample of firms 
from the larger Monthly Retail Trade Survey.
}
\keyword{datasets}
