% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate.tsissm.spec}
\alias{estimate.tsissm.spec}
\alias{estimate}
\title{Model Estimation}
\usage{
\method{estimate}{tsissm.spec}(
  object,
  solver = "nlminb",
  control = list(trace = 0),
  autodiff = TRUE,
  ...
)
}
\arguments{
\item{object}{an object of class \dQuote{tsissm.spec}.}

\item{solver}{one of either \dQuote{solnp}, \dQuote{nlminb} or \dQuote{optim}.
The latter uses the L-BFGS-B algorithm from the lbfgsb3c package. For option
\dQuote{autodiff}, valid solvers are \dQuote{nlminb} and \dQuote{nloptr}.}

\item{control}{solver control parameters.}

\item{autodiff}{whether to use automatic differentiation for estimation.
This makes use of the tsissmad package.}

\item{...}{only additional argument which can be passed when choosing autodiff
is that of \dQuote{use_hessian} which tells the solver to make use of second
derivatives.}
}
\value{
An object of class \dQuote{tsissm.estimate}
}
\description{
Estimates a model given a specification object using
maximum likelihood.
}
\details{
The maximum likelihood estimation for this model is described in the 
online book of tsmodels.
}
