\name{fitted}
\alias{fitted}
\alias{fitted.tsissm.estimate}
\title{
Model Fitted Values
}
\description{
Extract the fitted values from an estimated model.
}
\usage{
\method{fitted}{tsissm.estimate}(object, ...)
}
\arguments{
\item{object}{
An object of class tsissm.estimate
}
\item{\dots}{
Not currently used
}
}
\value{
An xts vector, back adjusted if a transformation was used.
}
\author{
Alexios Galanos
}
