\name{plot}
\alias{plot}
\alias{plot.tsissm.estimate}
\title{
Estimated Model Plots
}
\description{
Plots for objects generated from the tsissm functions.
}
\usage{
\method{plot}{tsissm.estimate}(x, y = NULL, ...)
}
\arguments{
\item{x}{
An object of class \dQuote{tsissm.estimate}.
}
\item{y}{
Not used.
}
\item{\dots}{
Additional arguments passed to the underlying plot function.
}
}
\author{
Alexios Galanos
}
