% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate.tsissm.estimate}
\alias{simulate.tsissm.estimate}
\alias{simulate}
\title{Model Simulation}
\usage{
\method{simulate}{tsissm.estimate}(
  object,
  nsim = 1,
  seed = NULL,
  h = NULL,
  newxreg = NULL,
  sim_dates = NULL,
  bootstrap = FALSE,
  innov = NULL,
  sigma_scale = 1,
  pars = coef(object),
  init_states = object$spec$xseed,
  ...
)
}
\arguments{
\item{object}{an object of class \dQuote{tsissm.estimate}.}

\item{nsim}{the number of paths per complete set of time steps (h).}

\item{seed}{an object specifying if and how the random number generator should be
initialized (\sQuote{seeded}). Either NULL or an integer that will be used in
a call to set.seed before simulating the response vectors. If set, the value
is saved as the "seed" attribute of the returned value. The default, NULL
will not change the random generator state, and return .Random.seed as the
\dQuote{seed} attribute in the returned object.}

\item{h}{the number of time steps to simulate paths for. If this is NULL,
it will use the same number of periods as in the original series.}

\item{newxreg}{an optional matrix of regressors to use for the simulation
if xreg was used in the estimation. If NULL and the estimated object had
regressors, and h was also set to NULL, then the original regressors will
be used.}

\item{sim_dates}{an optional vector of simulation dates equal to h. If NULL
will use the implied periodicity of the data to generate a regular sequence
of dates after the first available date in the data.}

\item{bootstrap}{whether to bootstrap the innovations from the estimated
object by re-sampliag from the empirical distribution.}

\item{innov}{an optional vector of uniform innovations which will be
translated to regular innovations using the appropriate distribution quantile
function and model standard deviation. The length of this vector should be
equal to nsim x horizon.}

\item{sigma_scale}{An optional scalar which will scale the standard deviation 
of the innovations (useful for profiling under different assumptions).}

\item{pars}{an optional named vector of model coefficients which override the
estimated coefficients. No checking is currently performed on the adequacy of
these coefficients.}

\item{init_states}{An optional vector of states to initialize the forecast. 
If NULL, will use the first available states from the estimated model.}

\item{...}{not currently used.}
}
\value{
An object of class \dQuote{tsissm.simulate} with slots for the simulated
series and states.
}
\description{
Simulation function for class \dQuote{tsissm.estimate}.
}
