% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.tsissm.estimate}
\alias{summary.tsissm.estimate}
\alias{summary}
\title{Model Estimation Summary}
\usage{
\method{summary}{tsissm.estimate}(object, digits = 4, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsissm.estimate}.}

\item{digits}{integer, used for number formatting. Optionally, to avoid
scientific notation, set \sQuote{options(scipen=999)}.}

\item{...}{not currently used.}
}
\value{
A printout of the parameter summary, model type and some model metrics.
When estimated using autodiff, the standard errors, t-values and p-values will
also be printed. In this case, if the parameters are close to their upper or
lower bounds then it is very likely that these values will be NaN.
}
\description{
Summary method for class \dQuote{tsissm.estimate}
}
