\name{summary}
\alias{summary}
\alias{summary.tsissm.estimate}
\title{
Model Estimation Summary
}
\description{
Summary method for class \dQuote{tsissm.estimate}
}
\usage{
\method{summary}{tsissm.estimate}(object, digits = 4, ...)
}
\arguments{
\item{object}{
An object of class tsissm.estimate
}
\item{digits}{
Integer, used for number formatting. Optionally, to avoid scientific notation,
set \sQuote{options(scipen=999)}.
}
\item{\dots}{
Not currently used
}
}
\value{
A printout of the parameter summary, model type and some model metrics.
}
\author{
Alexios Galanos
}
