% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{tsfilter.tsissm.estimate}
\alias{tsfilter.tsissm.estimate}
\alias{tsfilter}
\title{Online Model Filtering}
\usage{
\method{tsfilter}{tsissm.estimate}(object, y = NULL, newxreg = NULL, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsissm.estimate}.}

\item{y}{an xts vector of new information related to y. The function checks
whether y contains indices (dates) which are not in the passed object and
only filters new information.}

\item{newxreg}{An xts matrix of new information related to external regressors
(if those were used in the original model estimated).}

\item{...}{not currently used.}
}
\value{
An object of class \dQuote{tsissm.estimate}.
}
\description{
Online filter which updates the states and fitted values using
new data.
}
\details{
The new data is filtered (1 step ahead) using the last state of the
object. Once this is complete, the object is updated with the new states
and information so that the process can be continued on the same object as
new information arrives.
}
