% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsissm-package.R
\docType{package}
\name{tsissm-package}
\alias{tsissm}
\alias{tsissm-package}
\title{tsissm: Linear Innovations State Space Unobserved Components Model}
\description{
Unobserved Components Model using the Linear Innovation State Space representation.
}
\author{
\strong{Maintainer}: Alexios Galanos \email{alexios@4dscape.com}

Other contributors:
\itemize{
  \item James Nesbit \email{jmcgnesbit@gmail.com} [contributor]
  \item Keith O'Hara \email{kth.ohr@gmail.com} [contributor]
}

}
\keyword{internal}
