% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{tsmoments.tsissm.estimate}
\alias{tsmoments.tsissm.estimate}
\alias{tsmoments}
\title{Analytic Forecast Moments}
\usage{
\method{tsmoments}{tsissm.estimate}(
  object,
  h = 1,
  newxreg = NULL,
  init_states = NULL,
  transform = FALSE,
  ...
)
}
\arguments{
\item{object}{an object of class \dQuote{tsissm.estimate}.}

\item{h}{the forecast horizon.}

\item{newxreg}{a matrix of external regressors in the forecast horizon.}

\item{init_states}{optional vector of states to initialize the forecast.
If NULL, will use the last available state from the estimated model.}

\item{transform}{whether to backtransform the mean forecast. For the Box-Cox 
and logit transformations this uses a Taylor Series expansion to adjust for 
the bias.}

\item{...}{not currently used.}
}
\description{
Prediction function for class \dQuote{tsissm.estimate}.
}
